Content & Banner Security Update

+++ protected/controllers/back/BannerCategoryController.php
@@ -135,13 +135,17 @@ class BannerCategoryController extends BackEndController {
 
         if (Yii::app()->request->isPostRequest) {
             // we only allow deletion via POST request
-            $this->loadModel($id)->delete();
+            //check if there any dependancy
+            $total_sub_category = BannerCategory::getNumberOfSubCategory($id);
+            $total_content = BannerCategory::getNumberOfContent($id);
+            if ($total_sub_category <= 0 && $total_content <= 0) {
+                $this->loadModel($id)->delete();
+            }
 
             // if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
             if (!isset($_GET['ajax']))
                 $this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
-        }
-        else
+        } else
             throw new CHttpException(400, 'Invalid request. Please do not repeat this request again.');
     }
